# Conflicts:
#	src/main/java/com/example/heart_field/entity/Help.java
#	target/classes/com/example/heart_field/constant/RegexPattern.class
#	target/classes/com/example/heart_field/controller/AdminController.class
#	target/classes/com/example/heart_field/controller/ConsultantController$1.class
#	target/classes/com/example/heart_field/controller/ConsultantController$2.class
#	target/classes/com/example/heart_field/controller/ConsultantController.class
#	target/classes/com/example/heart_field/controller/UserApi.class
#	target/classes/com/example/heart_field/controller/VisitorController.class
#	target/classes/com/example/heart_field/dto/UserLoginDTO$UserLoginDTOBuilder.class
#	target/classes/com/example/heart_field/dto/UserLoginDTO.class
#	target/classes/com/example/heart_field/dto/consultant/binding/SupervisorBindedDto.class
#	target/classes/com/example/heart_field/dto/consultant/binding/SupervisorBinding.class
#	target/classes/com/example/heart_field/dto/consultant/comment/CommentDto.class
#	target/classes/com/example/heart_field/dto/consultant/profile/AnyConsultantProfileDto.class
#	target/classes/com/example/heart_field/dto/consultant/profile/ConsultantProfileDto.class
#	target/classes/com/example/heart_field/entity/Chat.class
#	target/classes/com/example/heart_field/entity/Consultant.class
#	target/classes/com/example/heart_field/entity/Help.class
#	target/classes/com/example/heart_field/entity/Schedule.class
#	target/classes/com/example/heart_field/entity/Supervisor.class
#	target/classes/com/example/heart_field/entity/User.class
#	target/classes/com/example/heart_field/filter/AuthenticationInterceptor.class
#	target/classes/com/example/heart_field/param/UserLoginParam.class
#	target/classes/com/example/heart_field/service/AdminService.class
#	target/classes/com/example/heart_field/service/ChatService.class
#	target/classes/com/example/heart_field/service/HelpService.class
#	target/classes/com/example/heart_field/service/ScheduleService.class
#	target/classes/com/example/heart_field/service/SupervisorService.class
#	target/classes/com/example/heart_field/service/UserService.class
#	target/classes/com/example/heart_field/service/VisitorService.class
#	target/classes/com/example/heart_field/service/impl/ChatServiceImpl.class
#	target/classes/com/example/heart_field/service/impl/HelpServiceImpl.class
#	target/classes/com/example/heart_field/service/impl/ScheduleServiceImpl.class
#	target/classes/com/example/heart_field/service/impl/SupervisorServiceImpl.class
#	target/classes/com/example/heart_field/service/impl/UserServiceImpl.class
#	target/classes/com/example/heart_field/service/impl/VisitorServiceImpl.class
#	target/classes/com/example/heart_field/tokens/TokenService.class
#	target/classes/com/example/heart_field/utils/CheckUtil.class
#	target/classes/com/example/heart_field/utils/TokenUtil.class
#	target/classes/com/example/heart_field/utils/UserUtils.class

+++ src/main/java/com/example/heart_field/dto/consultant/record/RecordListDTO.java
@@ -41,10 +41,10 @@ public class RecordListDTO {
     private LocalDateTime endTime;
 
     //咨询是否完成，0表示未完成，1表示已完成
-    private Byte isCompleted;
+    private Integer isCompleted;
 
     //访客对咨询师的评价
-    private Byte visitorScore;
+    private Integer visitorScore;
     private String visitorComment;
 
     //对应的聊天id和求助会话id
